services/diffusion/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
certifi==2022.12.7
charset-normalizer==3.1.0
diffusers==0.15.0
distlib==0.3.6
filelock==3.11.0
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.3.0
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.1
Pillow==9.5.0
pipenv==2023.3.20
platformdirs==3.2.0
psutil==5.9.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
safetensors==0.3.0
tokenizers==0.13.3
torch==1.13.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
zipp==3.15.0
